+++ checkpoint/CHANGELOG.md
@@ -7,8 +7,11 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [0.11.28] - 2025-11-06
+
 ### Added
 
+- #v1 `Checkpointer.wait` method.
 - Added `register_pathways_handlers` to `ocp.type_handlers`
 
 ## [0.11.27] - 2025-11-04